将容器类及相关接口的计数和索引类型从int32_t改为int64_t以支持更大数据量

+++ include/base/container/Array.h
@@ -9,7 +9,7 @@ namespace base
 	/// @brief 数组
 	///
 	///
-	template <typename ItemType, int32_t TCount>
+	template <typename ItemType, int64_t TCount>
 	class Array final :
 		public base::IRawArray<ItemType>
 	{
@@ -43,7 +43,7 @@ namespace base
 		///
 		Array(std::initializer_list<ItemType> const &list)
 		{
-			int32_t i = 0;
+			int64_t i = 0;
 			for (auto &item : list)
 			{
 				_arr[i++] = item;
@@ -67,7 +67,7 @@ namespace base
 		///
 		/// @return
 		///
-		virtual int32_t Count() const override
+		virtual int64_t Count() const override
 		{
 			return TCount;
 		}
