+++ include/base/container/Range.h
@@ -10,8 +10,8 @@ namespace base
 	class Range
 	{
 	private:
-		int32_t _begin = 0;
-		int32_t _end = 0;
+		int64_t _begin = 0;
+		int64_t _end = 0;
 
 	public:
 		///
@@ -20,7 +20,7 @@ namespace base
 		/// @param begin 区间左端点。是闭的。
 		/// @param end 区间右端点。是开的。
 		///
-		constexpr Range(int32_t begin, int32_t end)
+		constexpr Range(int64_t begin, int64_t end)
 			: _begin(begin),
 			  _end(end)
 		{
@@ -29,9 +29,9 @@ namespace base
 		///
 		/// @brief 区间左端点。区间左边是闭的。
 		///
-		/// @return int32_t
+		/// @return int64_t
 		///
-		constexpr int32_t Begin() const
+		constexpr int64_t Begin() const
 		{
 			return _begin;
 		}
@@ -39,9 +39,9 @@ namespace base
 		///
 		/// @brief 区间右端点。区间右边是开的。
 		///
-		/// @return int32_t
+		/// @return int64_t
 		///
-		constexpr int32_t End() const
+		constexpr int64_t End() const
 		{
 			return _end;
 		}
@@ -49,11 +49,12 @@ namespace base
 		///
 		/// @brief End - Begin，等于这个区间的大小。
 		///
-		/// @return int32_t
+		/// @return int64_t
 		///
-		constexpr int32_t Size() const
+		constexpr int64_t Size() const
 		{
 			return _end - _begin;
 		}
 	};
+
 } // namespace base
