+++ include/base/stream/StdOutStream.cpp
@@ -93,9 +93,9 @@ namespace
 		/// @brief 将本流的数据读取到 span 中。
 		///
 		/// @param span
-		/// @return int32_t
+		/// @return int64_t
 		///
-		virtual int32_t Read(base::Span const &span) override
+		virtual int64_t Read(base::Span const &span) override
 		{
 			throw base::NotSupportedException{};
 		}
